# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/v1/company":{"post":{"operationId":"CompanyController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new company","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":"company created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}},"tags":["Company"]},"get":{"operationId":"CompanyController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Company retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CompanyDto"}}}}}},"tags":["Company"]}},"/api/v1/company/{id}":{"get":{"operationId":"CompanyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Company retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}},"tags":["Company"]},"patch":{"operationId":"CompanyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":"Company Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}},"tags":["Company"]},"delete":{"operationId":"CompanyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Company deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Company"]}}},"info":{"title":"SUNBYTES","description":"Sunbytes Senior Backend Developer Task","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8989"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateCompanyDto":{"type":"object","properties":{"owner":{"type":"number"},"name":{"type":"string"},"company":{"type":"string"}},"required":["owner","name","company"]},"CompanyDto":{"type":"object","properties":{"id":{"type":"string","example":"hday334-ewui4-3whdsi-3836-dke0"},"owner":{"type":"number","example":12},"name":{"type":"string"},"company":{"type":"string"}},"required":["id","owner","name","company"]},"UpdateCompanyDto":{"type":"object","properties":{"owner":{"type":"number"},"name":{"type":"string"},"company":{"type":"string"}}},"StandardResponse":{"type":"object","properties":{"status":{"type":"string"},"code":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"},"meta":{"type":"object"}},"required":["status","code","message","data","meta"]}}}}